refactor: Move data nodes to individual files
Implements ML-597
